user  não encontrado delete

+++ src/controllers/user-controller.js
@@ -162,9 +162,19 @@ exports.put = async(req, res, next) => {
 };
 
 exports.delete = async(req, res, next) => {
+        
+    const user = await repository.getById(req.params.id);
+        
+    if (!user){
+        res.status(401).send({
+           mensagem: 'Usuário não encontrado'});            
+        return;
+    }
     
     try {
         await repository.delete(req.params.id)
+
+        
         res.status(200).send({
             mensagem: 'Usuário removido com sucesso!'
         });
